Extract abstract supperclass SensorMiniGameFragment
Cover light sensor minigame (optional) #64
