+++ app/src/main/java/com/se2/bopit/ui/games/CoverLightSensorMiniGame.java
@@ -12,12 +12,11 @@ import androidx.annotation.Nullable;
 
 import com.se2.bopit.R;
 import com.se2.bopit.domain.CoverLightSensorMiniGameModel;
-import com.se2.bopit.domain.SensorMiniGameModel;
 import com.se2.bopit.domain.annotations.RequireSensor;
-import com.se2.bopit.ui.MiniGameFragment;
+import com.se2.bopit.ui.SensorMiniGameFragment;
 
 @RequireSensor(CoverLightSensorMiniGameModel.SENSOR_TYPE)
-public class CoverLightSensorMiniGame extends MiniGameFragment<SensorMiniGameModel> {
+public class CoverLightSensorMiniGame extends SensorMiniGameFragment<CoverLightSensorMiniGameModel> {
 
     public CoverLightSensorMiniGame() {
         super(new CoverLightSensorMiniGameModel());
@@ -34,17 +33,5 @@ public class CoverLightSensorMiniGame extends MiniGameFragment<SensorMiniGameMod
         messageText.setText(R.string.challenge_cover_light_sensor);
 
         return view;
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        gameModel.resumeSensor(getContext());
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        gameModel.pauseSensor();
     }
 }
